requirements.txt
streamlit>=1.35.0
google-genai
python-dotenv
beautifulsoup4
requests